split grpc service from http service, error handling updates for both grpc and http
