+++ pkg/service/http_service.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	"github.com/open-feature/flagd/pkg/eval"
-	"github.com/open-feature/flagd/pkg/model"
 	gen "github.com/open-feature/flagd/schemas/protobuf/gen/v1"
 	log "github.com/sirupsen/logrus"
 	"github.com/soheilhy/cmux"
@@ -17,7 +16,6 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/grpc/status"
-	"google.golang.org/protobuf/types/known/structpb"
 )
 
 type HTTPServiceConfiguration struct {
@@ -26,14 +24,13 @@ type HTTPServiceConfiguration struct {
 
 type HTTPService struct {
 	HTTPServiceConfiguration *HTTPServiceConfiguration
-	eval                     eval.IEvaluator
-	gen.UnimplementedServiceServer
+	GRPCService              *GRPCService
 }
 
 func (s *HTTPService) Serve(ctx context.Context, eval eval.IEvaluator) error {
-	s.eval = eval
+	s.GRPCService.eval = eval
 	grpcServer := grpc.NewServer()
-	gen.RegisterServiceServer(grpcServer, s)
+	gen.RegisterServiceServer(grpcServer, s.GRPCService)
 
 	mux := runtime.NewServeMux(
 		runtime.WithErrorHandler(s.HTTPErrorHandler),
@@ -77,115 +74,31 @@ func (s HTTPService) HTTPErrorHandler(
 	r *http.Request,
 	err error,
 ) {
-	if s, ok := status.FromError(err); ok {
-		code := s.Code()
-		switch {
-		case codes.Unknown == code:
-			w.WriteHeader(http.StatusInternalServerError)
-		case codes.InvalidArgument == code:
-			w.WriteHeader(http.StatusBadRequest)
-		case codes.NotFound == code:
-			w.WriteHeader(http.StatusNotFound)
-		default:
-			w.WriteHeader(http.StatusInternalServerError)
-		}
-		var res []byte
-		if res, err = json.Marshal(gen.ErrorResponse{
-			ErrorCode: s.Message(),
-		}); err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			log.Error(err)
-			return
-		}
-		if _, err = w.Write(res); err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			log.Error(err)
-			return
-		}
-	} else {
+	st := status.Convert(err)
+	switch {
+	case st.Code() == codes.Unknown:
 		w.WriteHeader(http.StatusInternalServerError)
+	case st.Code() == codes.InvalidArgument:
+		w.WriteHeader(http.StatusBadRequest)
+	case st.Code() == codes.NotFound:
+		w.WriteHeader(http.StatusNotFound)
+	default:
+		w.WriteHeader(http.StatusInternalServerError)
+	}
+	details := st.Details()
+	if len(details) != 1 {
+		log.Errorf("malformed error recieved by error handler, details recieved: %d - %v", len(details), details)
+		return
+	}
+	var res []byte
+	if res, err = json.Marshal(details[0]); err != nil {
 		log.Error(err)
+		return
 	}
-}
-
-// TODO: might be able to simplify some of this with generics.
-func (s HTTPService) ResolveBoolean(
-	ctx context.Context,
-	req *gen.ResolveBooleanRequest,
-) (*gen.ResolveBooleanResponse, error) {
-	res := gen.ResolveBooleanResponse{}
-	result, variant, reason, err := s.eval.ResolveBooleanValue(req.GetFlagKey(), req.GetContext())
-	if err != nil {
-		return &res, handleEvaluationError(err, reason)
+	if _, err = w.Write(res); err != nil {
+		log.Error(err)
+		return
 	}
-	res.Reason = reason
-	res.Value = result
-	res.Variant = variant
-	return &res, nil
-}
-
-func (s HTTPService) ResolveString(
-	ctx context.Context,
-	req *gen.ResolveStringRequest,
-) (*gen.ResolveStringResponse, error) {
-	res := gen.ResolveStringResponse{}
-	result, variant, reason, err := s.eval.ResolveStringValue(req.GetFlagKey(), req.GetContext())
-	if err != nil {
-		return &res, handleEvaluationError(err, reason)
-	}
-	res.Reason = reason
-	res.Value = result
-	res.Variant = variant
-	return &res, nil
-}
-
-func (s HTTPService) ResolveNumber(
-	ctx context.Context,
-	req *gen.ResolveNumberRequest,
-) (*gen.ResolveNumberResponse, error) {
-	res := gen.ResolveNumberResponse{}
-	result, variant, reason, err := s.eval.ResolveNumberValue(req.GetFlagKey(), req.GetContext())
-	if err != nil {
-		return &res, handleEvaluationError(err, reason)
-	}
-	res.Reason = reason
-	res.Value = result
-	res.Variant = variant
-	return &res, nil
-}
-
-func (s HTTPService) ResolveObject(
-	ctx context.Context,
-	req *gen.ResolveObjectRequest,
-) (*gen.ResolveObjectResponse, error) {
-	res := gen.ResolveObjectResponse{}
-	result, variant, reason, err := s.eval.ResolveObjectValue(req.GetFlagKey(), req.GetContext())
-	if err != nil {
-		return &res, handleEvaluationError(err, reason)
-	}
-	val, err := structpb.NewStruct(result)
-	if err != nil {
-		return &res, handleEvaluationError(err, reason)
-	}
-	res.Reason = reason
-	res.Value = val
-	res.Variant = variant
-	return &res, nil
-}
-
-// some basic mapping of errors from model to HTTP
-func handleEvaluationError(err error, reason string) error {
-	// TODO: we should consider creating a custom error that includes a code instead of using the message for this.
-	statusCode := codes.Internal
-	message := err.Error()
-	switch message {
-	case model.FlagNotFoundErrorCode:
-		statusCode = codes.NotFound
-	case model.TypeMismatchErrorCode:
-		statusCode = codes.InvalidArgument
-	}
-	log.Error(message)
-	return status.Error(statusCode, message)
 }
 
 // TODO: could be replaced with a logging client
